export with equal sign refactored
