--- sources/system/print_2d_array_fd.c
@@ -1,25 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   print_2d_array_fd.c                                :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: lcouto <lcouto@student.42.fr>              +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/07/11 19:51:58 by lcouto            #+#    #+#             */
-/*   Updated: 2021/07/11 20:09:38 by lcouto           ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-void	print_2d_array_fd(char **array, int fd)
-{
-	int	i;
-
-	i = 0;
-	while (array[i])
-	{
-		ft_putendl_fd(array[i], fd);
-		i++;
-	}
-}
